### 2024-08-30 #7161 @editorialbot
[PRE REVIEW]: pyuvdata v3: an interface for astronomical interferometeric datasets in python
**Submitting author:** @kartographer (<a href="http://orcid.org/0000-0002-3490-146X">Garrett Keating</a>)
**Repository:** https://github.com/RadioAstronomySoftwareGroup/pyuvdata/
**Branch with paper.md** (empty if default branch): joss_update
**Version:** v3.0
