### 2024-09-01 #7161 @kartographer
Hi @dfm!

> In the meantime, since this submission is specifically about the updates to this project, could you add a comment here with more details about exactly what should be considered part of this review?

There's actually a section on this in the submitted paper, but in brief, since our [original JOSS paper](https://joss.theoj.org/papers/10.21105/joss.00140) the package has undergone a pretty substantial expansion in capabilities, as three of the four main modules (`UVBeam`, `UVCal`, and `UVFlag`) are new to the package, and the last one (`UVData`) has been fairly extensively re-written to expand the capabilities of the package. As such, I don't think there's anything that I would consider "out of scope". Hopefully that addresses your question here, but let me know if there's some more detail you're looking for here.

> It would also be useful to include a link to the commit history for this version so that we can also see what changed explicitly in the code.

Commit history is [here](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/commits/main/), tag for the relevant release is [here](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/releases/tag/v3.0.0), and tag for the previously published release is [here](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/releases/tag/v1.0).
