+++ .pre-commit-config.yaml
@@ -17,7 +17,7 @@ repos:
           args: ['--fix=no']
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.6.2
+    rev: v0.6.3
     hooks:
       - id: ruff
       - id: ruff-format
