### 2024-09-03 fc5aad9
Forcing defensive approach w/ text fields inside of codes_read (due to older data)
