### 2024-09-03 622ce55
Forcing defensive approach w/ text fields inside of codes_read (due to older data)
