### 2024-09-03 d1279b4
Adding catch for key errors when populating ac_data structure
