### 2024-09-03 38cdc87
Adding catch for key errors when populating ac_data structure
