### 2024-09-04 c572404
refactor: move regularity check to check()
