### 2024-09-04 ebf5cd4
test: fix tests that errored on non-even axes
