### 2024-09-04 2221851
test: fix tests that errored on non-even axes
