### 2024-09-04 ffd28cf
error rather than warn in select for unevenly spaced axes
Also update to use the utils method for checking array spacing
