### 2024-09-04 3720f46
move for loop in test to parametrize
