### 2024-09-06 #7161 @JBorrow
La Planete is mis-spelled in the references as far as I can see.
