### 2024-09-08 d80eb04
use dataclass for Analytic beams and Beam Interface objects
