### 2024-09-08 eb10fe1
use dataclass for Analytic beams and Beam Interface objects
