+++ src/pyuvdata/uvbeam/analytic_beam.py
@@ -6,7 +6,8 @@
 from __future__ import annotations
 
 from abc import ABC, abstractmethod
-from typing import Any, Literal
+from dataclasses import dataclass
+from typing import Literal
 
 import numpy as np
 import numpy.typing as npt
@@ -20,6 +21,7 @@ from .uvbeam import UVBeam, _convert_feeds_to_pols
 __all__ = ["AnalyticBeam", "AiryBeam", "GaussianBeam", "ShortDipoleBeam", "UniformBeam"]
 
 
+@dataclass
 class AnalyticBeam(ABC):
     """
     Define an analytic beam base class.
@@ -55,17 +57,12 @@ class AnalyticBeam(ABC):
 
     # In the future, this might allow for cartesian basis vectors in some orientation.
     # In that case, the Naxes_vec would be 3 rather than 2
-    basis_vec_dict = {"az_za": 2}
+    _basis_vec_dict = {"az_za": 2}
 
     @property
     @abstractmethod
     def basis_vector_type(self):
         """Require that a basis_vector_type is defined in concrete classes."""
-
-    @property
-    @abstractmethod
-    def name(self):
-        """Require that a name is defined in concrete classes."""
 
     def __init__(
         self,
@@ -74,12 +71,12 @@ class AnalyticBeam(ABC):
         include_cross_pols: bool = True,
         x_orientation: Literal["east", "north"] = "east",
     ):
-        if self.basis_vector_type not in self.basis_vec_dict:
+        if self.basis_vector_type not in self._basis_vec_dict:
             raise ValueError(
                 f"basis_vector_type is {self.basis_vector_type}, must be one of "
-                f"{list(self.basis_vec_dict.keys())}"
+                f"{list(self._basis_vec_dict.keys())}"
             )
-        self.Naxes_vec = self.basis_vec_dict[self.basis_vector_type]
+        self.Naxes_vec = self._basis_vec_dict[self.basis_vector_type]
 
         if feed_array is not None:
             for feed in feed_array:
@@ -97,111 +94,6 @@ class AnalyticBeam(ABC):
             self.feed_array, include_cross_pols, x_orientation=self.x_orientation
         )
         self.Npols = self.polarization_array.size
-
-    def __iter__(self) -> str:
-        """
-        Iterate over all attributes.
-
-        Yields
-        ------
-        attribute
-            Object attributes.
-
-        """
-        attribute_list = [
-            a
-            for a in dir(self)
-            if not a.startswith("_") and not callable(getattr(self, a))
-        ]
-        yield from attribute_list
-
-    def __eq__(self, other: Any, silent: bool = False) -> bool:
-        """Equality method."""
-        if isinstance(other, self.__class__):
-            # First check that the set of attrs is the same
-            self_attrs = set(self.__iter__())
-            other_attrs = set(other.__iter__())
-            if self_attrs != other_attrs:
-                if not silent:
-                    print(
-                        "Sets of parameters do not match. "
-                        f"Left is {self_attrs},"
-                        f" right is {other_attrs}. Left has "
-                        f"{self_attrs.difference(other_attrs)} extra."
-                        f" Right has {other_attrs.difference(self_attrs)} extra."
-                    )
-                return False
-
-            # Now check that attributes are equal
-            a_equal = True
-            for attr in self_attrs:
-                self_attr = getattr(self, attr)
-                other_attr = getattr(other, attr)
-                if not isinstance(other_attr, self_attr.__class__):
-                    if not silent:
-                        print(
-                            f"attribute {attr} are not the same types. "
-                            f"Left is {type(self_attr)}, right is {type(other_attr)}."
-                        )
-                    a_equal = False
-                elif isinstance(self_attr, np.ndarray):
[…]
