### 2024-09-09 b73180b
correctly pass spline_opts to new uvbeam interpolator
