+++ paper.bib
@@ -98,7 +98,7 @@
 }
 
 @TECHREPORT{calfits,
-  author = {{Ali}, Z. and  {Hazelton}, B. and {Beardsley}, A. and {La Planete}, P. and {Kunicki}, T. and {the pyuvdata team}},
+  author = {{Ali}, Z. and  {Hazelton}, B. and {Beardsley}, A. and {La Plante}, P. and {Kunicki}, T. and {the pyuvdata team}},
   title = "{Memo: UVCal FITS Format}",
   year = 2017,
   month = jul,
