### 2024-09-13 a1ffbd5
Better inputs and docstring for AnalyticBeam.to_uvbeam
