### 2024-09-13 37179a6
Better inputs and docstring for AnalyticBeam.to_uvbeam
