### 2024-09-17 ce7c884
Add yaml constructors and representers for Analytic and UV beams
