+++ src/pyuvdata/uvbeam/analytic_beam.py
@@ -5,12 +5,15 @@
 
 from __future__ import annotations
 
+import dataclasses
+import importlib
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
 from typing import Literal
 
 import numpy as np
 import numpy.typing as npt
+import yaml
 from astropy.constants import c as speed_of_light
 from scipy.special import j1
 
@@ -18,7 +21,8 @@ from .. import utils
 from ..docstrings import combine_docstrings
 from .uvbeam import UVBeam, _convert_feeds_to_pols
 
-__all__ = ["AnalyticBeam", "AiryBeam", "GaussianBeam", "ShortDipoleBeam", "UniformBeam"]
+analytic_beam_classes = ["AiryBeam", "GaussianBeam", "ShortDipoleBeam", "UniformBeam"]
+__all__ = ["AnalyticBeam"] + analytic_beam_classes
 
 
 @dataclass
@@ -67,7 +71,7 @@ class AnalyticBeam(ABC):
     def __init__(
         self,
         *,
-        feed_array: npt.NDArray[np.str] | None = None,
+        feed_array: npt.NDArray[str] | None = None,
         include_cross_pols: bool = True,
         x_orientation: Literal["east", "north"] = "east",
     ):
@@ -98,9 +102,9 @@ class AnalyticBeam(ABC):
     def _check_eval_inputs(
         self,
         *,
-        az_array: npt.NDArray[np.float],
-        za_array: npt.NDArray[np.float],
-        freq_array: npt.NDArray[np.float],
+        az_array: npt.NDArray[float],
+        za_array: npt.NDArray[float],
+        freq_array: npt.NDArray[float],
     ):
         """Check the inputs for the eval methods."""
         if az_array.ndim > 1 or za_array.ndim > 1 or freq_array.ndim > 1:
@@ -113,7 +117,7 @@ class AnalyticBeam(ABC):
 
     def _get_empty_data_array(
         self, npts: int, nfreqs: int, beam_type: str = "efield"
-    ) -> npt.NDArray[np.float]:
+    ) -> npt.NDArray[float]:
         """Get the empty data to fill in the eval methods."""
         if beam_type == "efield":
             return np.zeros((self.Naxes_vec, self.Nfeeds, nfreqs, npts), dtype=complex)
@@ -129,10 +133,10 @@ class AnalyticBeam(ABC):
     def _efield_eval(
         self,
         *,
-        az_array: npt.NDArray[np.float],
-        za_array: npt.NDArray[np.float],
-        freq_array: npt.NDArray[np.float],
-    ) -> npt.NDArray[np.float]:
+        az_array: npt.NDArray[float],
+        za_array: npt.NDArray[float],
+        freq_array: npt.NDArray[float],
+    ) -> npt.NDArray[float]:
         """
         Evaluate the efield at the given coordinates.
 
@@ -161,10 +165,10 @@ class AnalyticBeam(ABC):
     def efield_eval(
         self,
         *,
-        az_array: npt.NDArray[np.float],
-        za_array: npt.NDArray[np.float],
-        freq_array: npt.NDArray[np.float],
-    ) -> npt.NDArray[np.float]:
+        az_array: npt.NDArray[float],
+        za_array: npt.NDArray[float],
+        freq_array: npt.NDArray[float],
+    ) -> npt.NDArray[float]:
         """
         Evaluate the efield at the given coordinates.
 
@@ -198,10 +202,10 @@ class AnalyticBeam(ABC):
     def _power_eval(
         self,
         *,
-        az_array: npt.NDArray[np.float],
-        za_array: npt.NDArray[np.float],
-        freq_array: npt.NDArray[np.float],
[…]
