### 2024-09-17 3bd7312
Add yaml constructors and representers for Analytic and UV beams
