### 2024-09-17 1eb09a6
Fix some things in UVBeam yaml handling, add test coverage
