### 2024-09-17 d962e1a
Fix some things in UVBeam yaml handling, add test coverage
