### 2024-09-17 c18da8f
remove comment and comment on alert. We only really want summary
