### 2024-09-17 849a550
remove comment and comment on alert. We only really want summary
