### 2024-09-17 53a12a5
remove unnecessary pytest-benchmark options; restore comment always
options don't seem to affect the results
