+++ .github/workflows/macosx_windows_ci.yaml
@@ -129,7 +129,7 @@ jobs:
 
       - name: Run benchmark
         run: |
-          pytest --benchmark-only --benchmark-name=short --benchmark-json output.json
+          pytest --benchmark-only --benchmark-json output.json
 
       # Download previous benchmark result from cache (if exists)
       - name: Download previous benchmark data
@@ -150,6 +150,10 @@ jobs:
           external-data-json-path: ./cache/benchmark-data.json
           # Workflow will fail when an alert happens
           fail-on-alert: true
+          # Enable commenting
+          comment-always: true
+          # Comment on the PR if the branch is not a fork
+          comment-on-alert: true
           # Enable Job Summary for PRs
           summary-always: true
           github-token: ${{ secrets.GITHUB_TOKEN }}
