### 2024-09-17 efda0e3
remove unnecessary pytest-benchmark options; restore comment always
options don't seem to affect the results
