+++ .github/workflows/macosx_windows_ci.yaml
@@ -150,8 +150,6 @@ jobs:
           external-data-json-path: ./cache/benchmark-data.json
           # Workflow will fail when an alert happens
           fail-on-alert: true
-          # Enable commenting
-          comment-always: true
           # Comment on the PR if the branch is not a fork
           comment-on-alert: true
           # Enable Job Summary for PRs
