### 2024-09-17 ada6d64
better dataclass setup, other PR comment fixes
remove accidentally checked in plugin update
