### 2024-09-17 2792c37
better dataclass setup, other PR comment fixes
remove accidentally checked in plugin update
