+++ src/pyuvdata/uvbeam/analytic_beam.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import dataclasses
 import importlib
 from abc import ABC, abstractmethod
-from dataclasses import dataclass
+from dataclasses import InitVar, dataclass, field
 from typing import Literal
 
 import numpy as np
@@ -21,43 +21,44 @@ from .. import utils
 from ..docstrings import combine_docstrings
 from .uvbeam import UVBeam, _convert_feeds_to_pols
 
-analytic_beam_classes = ["AiryBeam", "GaussianBeam", "ShortDipoleBeam", "UniformBeam"]
-__all__ = ["AnalyticBeam"] + analytic_beam_classes
+__all__ = ["AnalyticBeam", "AiryBeam", "GaussianBeam", "ShortDipoleBeam", "UniformBeam"]
 
 
-@dataclass
+@dataclass(kw_only=True)
 class AnalyticBeam(ABC):
     """
     Define an analytic beam base class.
 
     Attributes
     ----------
-    basis_vec_dict : dict
-        Allowed basis vector types (keys) with their number of axes (which will be set
-        on the object as Naxes_vec).
-    basis_vector_type : str
-        The type of basis vectors this beam is defined with. Must be a key in
-        `basis_vec_dict`. Currently, only "az_za" is supported, which means basis
-        vectors aligned with the zenith angle and azimuth.
-    Naxes_vec : int
-        Number of basis vectors. This is set based on the `basis_vector_type`.
-    Nfeeds : int
-        Number of feeds. This is set based on the `feed_array`.
+    feed_array : np.ndarray of str
+        Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
+        or r & l.
+    x_orientation : str
+        Physical orientation of the feed for the x feed. Not meaningful for
+        unpolarized analytic beams, but clarifies the orientation of the dipole
+        for the ShortDipoleBeam and matches with the meaning on UVBeam objects.
 
     Parameters
     ----------
     feed_array : np.ndarray of str
         Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
         or r & l.
-    include_cross_pols : bool
-        Option to include the cross polarized beams (e.g. xy and yx or en and ne) for
-        the power beam.
     x_orientation : str
         Physical orientation of the feed for the x feed. Not meaningful for
         unpolarized analytic beams, but clarifies the orientation of the dipole
         for the ShortDipoleBeam and matches with the meaning on UVBeam objects.
+    include_cross_pols : bool
+        Option to include the cross polarized beams (e.g. xy and yx or en and ne) for
+        the power beam.
 
     """
+
+    feed_array: npt.NDArray[str] | None = field(default=None, repr=False, compare=False)
+    x_orientation: Literal["east", "north"] = field(
+        default="east", repr=False, compare=False
+    )
+    include_cross_pols: InitVar[bool] = True
 
     # In the future, this might allow for cartesian basis vectors in some orientation.
     # In that case, the Naxes_vec would be 3 rather than 2
@@ -68,36 +69,50 @@ class AnalyticBeam(ABC):
     def basis_vector_type(self):
         """Require that a basis_vector_type is defined in concrete classes."""
 
-    def __init__(
-        self,
-        *,
-        feed_array: npt.NDArray[str] | None = None,
-        include_cross_pols: bool = True,
-        x_orientation: Literal["east", "north"] = "east",
-    ):
+    def __post_init__(self, include_cross_pols):
+        """
+        Post-initialization validation and conversions.
+
+        Parameters
+        ----------
+        include_cross_pols : bool
+            Option to include the cross polarized beams (e.g. xy and yx or en and ne)
+            for the power beam.
+
+        """
[…]
