### 2024-09-17 f2c3e2f
Add plugin infrastructure
