### 2024-09-17 5e493e3
Add plugin infrastructure
