### 2024-09-17 cd03053
Move subclass check into `__init_subclass__`
