### 2024-09-17 200a86c
Move subclass check into `__init_subclass__`
