### 2024-09-18 f126a9a
CI fix: Mambaforge is deprecated and mamba in now the default installer
