### 2024-09-18 bcc7bb8
move analytic_beam.py and beam_interface.py to top level
