### 2024-09-18 af8ccd4
move analytic_beam.py and beam_interface.py to top level
