+++ src/pyuvdata/analytic_beam.py
@@ -27,7 +27,7 @@ __all__ = ["AnalyticBeam", "AiryBeam", "GaussianBeam", "ShortDipoleBeam", "Unifo
 @dataclass(kw_only=True)
 class AnalyticBeam(ABC):
     """
-    Define an analytic beam base class.
+    Analytic beam base class.
 
     Attributes
     ----------
@@ -37,7 +37,8 @@ class AnalyticBeam(ABC):
     x_orientation : str
         Physical orientation of the feed for the x feed. Not meaningful for
         unpolarized analytic beams, but clarifies the orientation of the dipole
-        for the ShortDipoleBeam and matches with the meaning on UVBeam objects.
+        for polarized beams like the ShortDipoleBeam and matches with the meaning
+        on UVBeam objects.
 
     Parameters
     ----------
@@ -47,7 +48,8 @@ class AnalyticBeam(ABC):
     x_orientation : str
         Physical orientation of the feed for the x feed. Not meaningful for
         unpolarized analytic beams, but clarifies the orientation of the dipole
-        for the ShortDipoleBeam and matches with the meaning on UVBeam objects.
+        for for polarized beams like the ShortDipoleBeam and matches with the
+        meaning on UVBeam objects
     include_cross_pols : bool
         Option to include the cross polarized beams (e.g. xy and yx or en and ne) for
         the power beam.
@@ -529,6 +531,114 @@ yaml.add_multi_representer(
 yaml.add_multi_representer(AnalyticBeam, _analytic_beam_representer, Dumper=yaml.Dumper)
 
 
+@dataclass(kw_only=True)
+class AiryBeam(AnalyticBeam):
+    """
+    A zenith pointed Airy beam.
+
+    Airy beams are the diffraction pattern of a circular aperture, so represent
+    an idealized dish. Requires a dish diameter in meters and is inherently
+    chromatic and unpolarized.
+
+    The unpolarized nature leads to some results that may be surprising to radio
+    astronomers: if two feeds are specified they will have identical responses
+    and the cross power beam between the two feeds will be identical to the
+    spower beam for a single feed.
+
+    Attributes
+    ----------
+    diameter : float
+        Dish diameter in meters.
+    feed_array : np.ndarray of str
+        Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
+        or r & l.
+    x_orientation : str
+        Physical orientation of the feed for the x feed. Not meaningful for
+        AiryBeams, which are unpolarized.
+
+    Parameters
+    ----------
+    diameter : float
+        Dish diameter in meters.
+    feed_array : np.ndarray of str
+        Feeds to define this beam for, e.g. n & e or x & y or r & l.
+    x_orientation : str
+        Physical orientation of the feed for the x feed. Not meaningful for
+        AiryBeams, which are unpolarized.
+    include_cross_pols : bool
+        Option to include the cross polarized beams (e.g. xy and yx or en and ne) for
+        the power beam.
+
+    """
+
+    diameter: float
+    feed_array: npt.NDArray[str] | None = field(default=None, repr=False, compare=False)
+    x_orientation: Literal["east", "north"] = field(
+        default="east", repr=False, compare=False
+    )
+
+    include_cross_pols: InitVar[bool] = True
+
+    basis_vector_type = "az_za"
+
+    def _efield_eval(
+        self,
+        *,
+        az_array: npt.NDArray[float],
+        za_array: npt.NDArray[float],
+        freq_array: npt.NDArray[float],
+    ) -> npt.NDArray[float]:
+        """Evaluate the efield at the given coordinates."""
+        data_array = self._get_empty_data_array(az_array.size, freq_array.size)
+
[…]
