### 2024-09-24 13eea2c
add mwa and mwax reading benchmarks
