### 2024-09-24 e3f9e78
Add a tutorial on analytic beams
