+++ docs/tutorial.rst
@@ -16,3 +16,4 @@ Tutorials are available for each major user class:
     uvdata_tutorial.rst
     uvcal_tutorial.rst
     uvbeam_tutorial.rst
+    analytic_beam_tutorial.rst
