### 2024-09-24 650526e
Add a tutorial on analytic beams
