+++ docs/analytic_beam_tutorial.rst
@@ -1,0 +1,753 @@
+.. _analytic_beam_tutorial:
+
+--------------
+Analytic Beams
+--------------
+
+The analytic beams defined in pyuvdata are based on an abstract base class,
+:class:`pyuvdata.analytic_beam.AnalyticBeam`, which ensures a standard interface
+and can be used to define other analytic beams in a consistent way.
+
+
+Evaluating analytic beams
+-------------------------
+
+To evaluate an analytic beam at one or more frequencies and in in one or more
+directions, use either the :meth:`pyuvdata.analytic_beam.AnalyticBeam.efield_eval`
+or :meth:`pyuvdata.analytic_beam.AnalyticBeam.power_eval` methods as appropriate.
+
+Evaluating an Airy Beam power response
+**************************************
+
+This code evaluates and plots an Airy beam power response. Note that we exclude
+the cross polarizations, since this is an unpolarized the cross polarizations
+are identical to the auto polarization power beams. If the cross polarizations
+are included the array returned from the ``power_eval`` method will be complex.
+
+.. code-block:: python
+
+  >>> import matplotlib.pyplot as plt
+  >>> import numpy as np
+  >>> from matplotlib.colors import LogNorm
+
+  >>> from pyuvdata import AiryBeam
+
+  >>> # Create an AiryBeam with a diameter of 14.5 meters
+  >>> airy_beam = AiryBeam(diameter=14.5, include_cross_pols=False)
+
+  >>> # set up zenith angle, azimuth and frequency arrays to evaluate with
+  >>> # make a regular grid in direction cosines for nice plots
+  >>> n_vals = 100
+  >>> zmax = np.radians(90)  # Degrees
+  >>> axis_arr = np.arange(-n_vals/2., n_vals/2.) / float(n_vals/2.)
+  >>> l_arr, m_arr = np.meshgrid(axis_arr, axis_arr)
+  >>> radius = np.sqrt(l_arr**2 + m_arr**2)
+  >>> za_array = radius * zmax
+  >>> az_array = np.arctan2(m_arr, l_arr)
+
+  >>> Nfreqs = 11
+  >>> freqs = np.linspace(100, 200, 11) * 1e6
+
+  >>> beam_vals = airy_beam.power_eval(
+  ...     az_array=az_array.flatten(), za_array=za_array.flatten(), freq_array=freqs
+  ... )
+  >>> # Note that the returned array for the power response has shape
+  >>> # (1, Npols, number_of_frequencies, number_of_directions)
+  >>> beam_vals.shape
+  (1, 2, 11, 10000)
+
+  >>> beam_vals = np.reshape(beam_vals, (1, airy_beam.Npols, Nfreqs, n_vals, n_vals))
+
+  >>> fig, ax = plt.subplots(1, 2)
+  >>> bp_low = ax[0].imshow(
+  ...   beam_vals[0,0,0],
+  ...   norm=LogNorm(vmin = 1e-8, vmax =1),
+  ...   extent=[np.min(l_arr), np.max(l_arr), np.min(m_arr), np.max(m_arr)],
+  ... )
+  >>> _ = ax[0].set_title(f"Airy beam {freqs[0]*1e-6} MHz")
+  >>> _ = ax[0].set_xlabel("direction cosine l")
+  >>> _ = ax[0].set_ylabel("direction cosine m")
+  >>> _ = fig.colorbar(bp_low, ax=ax[0], fraction=0.046, pad=0.04)
+
+  >>> bp_high = ax[1].imshow(
+  ...   beam_vals[0,0,-1],
[…]
