### 2024-09-24 c913247
Now correctly references the Telescope object
