### 2024-09-24 98043a5
Now correctly references the Telescope object
