### 2024-09-24 592a48a
Now correctly references the Telescope object
