### 2024-09-30 330d315
Add a BeamInterface tutorial
minor improvements to the analytic beam tutorial
better error message in UVBeam.interp
