### 2024-09-30 8f3d7d9
Add a BeamInterface tutorial
minor improvements to the analytic beam tutorial
better error message in UVBeam.interp
