+++ docs/beam_interface.rst
@@ -3,11 +3,11 @@ Beam Interface
 
 The BeamInterface object is designed to provide a unified interface for UVBeam
 and AnalyticBeam objects to compute beam response values. It can be constructed
-with either a UVBeam or AnalyticBeam and it provides the
-:meth:`pyuvdata.BeamInterface.compute_response` method, which can be used to
+with either a :class:`pyuvdata.UVBeam` or :class:`AnalyticBeam` and it provides
+the :meth:`pyuvdata.BeamInterface.compute_response` method, which can be used to
 either evaluate the response of an AnalyticBeam or interpolate a UVBeam to get
 the beam response. This interface provides a simplified way for simulators to
-call either UVBeams or analytic beams.
+get beam responses from either UVBeams or analytic beams with the same code.
 
 .. autoclass:: pyuvdata.BeamInterface
    :members:
