+++ docs/tutorial.rst
@@ -17,3 +17,4 @@ Tutorials are available for each major user class:
     uvcal_tutorial.rst
     uvbeam_tutorial.rst
     analytic_beam_tutorial.rst
+    beam_interface_tutorial.rst
