### 2024-10-02 9455ec5
simplify unpolarized analytic beam construction
