+++ .pre-commit-config.yaml
@@ -2,7 +2,7 @@ exclude: '(^docs/conf.py|^docs/references/|^src/pyuvdata/data/|^src/pyuvdata/uvd
 
 repos:
   - repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v4.6.0
+    rev: v5.0.0
     hooks:
         - id: trailing-whitespace
         - id: check-added-large-files
@@ -17,7 +17,7 @@ repos:
           args: ['--fix=no']
 
   - repo: https://github.com/astral-sh/ruff-pre-commit
-    rev: v0.6.8
+    rev: v0.6.9
     hooks:
       - id: ruff
       - id: ruff-format
