### 2024-10-08 87f6fb3
added distance units and boolean near-field flag to the phase_dict object
