### 2024-10-10 d0694f5
perf: don't always return basis-vector array
