### 2024-10-10 e4878f1
perf: don't always return basis-vector array
