+++ CHANGELOG.md
@@ -38,6 +38,10 @@ antennas with data.
 - A bug in reading UVH5 files with antenna names saved as variable length strings
 that was introduced in v3.0.0.
 
+### Changed
+- Made it possible to *not* return the `interp_basis_vector` array from beam
+interpolators.
+
 ## [3.0.0] - 2024-7-1
 
 ### Added
