### 2024-10-10 50b3c89
test: new test of setting no basis vector return
