### 2024-10-10 c144d92
test: new test of setting no basis vector return
