### 2024-10-10 c6363d5
test: new test of setting no basis vector return
