### 2024-10-10 #7161 @JBorrow
Hi @@David-McKenna, I noticed you were an author of a similar package submitted to JOSS and I think you'd be a great reviewer for this paper. Would that be something you'd be interested in? Our reviewer's guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
