+++ src/pyuvdata/analytic_beam.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 import dataclasses
 import importlib
-from abc import ABC, abstractmethod
+import warnings
 from dataclasses import InitVar, dataclass, field
 from typing import ClassVar, Literal
 
@@ -25,13 +25,13 @@ __all__ = ["AnalyticBeam", "AiryBeam", "GaussianBeam", "ShortDipoleBeam", "Unifo
 
 
 @dataclass(kw_only=True)
-class AnalyticBeam(ABC):
+class AnalyticBeam:
     """
     Analytic beam base class.
 
     Attributes
     ----------
-    feed_array : np.ndarray of str
+    feeds : list of str
         Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
         or r & l.
     x_orientation : str
@@ -42,9 +42,9 @@ class AnalyticBeam(ABC):
 
     Parameters
     ----------
-    feed_array : np.ndarray of str
+    feeds : list of str
         Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
-        or r & l.
+        or r & l. Default is ["x", "y"].
     x_orientation : str
         Physical orientation of the feed for the x feed. Not meaningful for
         unpolarized analytic beams, but clarifies the orientation of the dipole
@@ -56,11 +56,12 @@ class AnalyticBeam(ABC):
 
     """
 
-    feed_array: npt.NDArray[str] | None = field(default=None, repr=False, compare=False)
-    x_orientation: Literal["east", "north"] = field(
-        default="east", repr=False, compare=False
-    )
+    feeds: list[str] | None = None
+    x_orientation: Literal["east", "north"] | None = "east"
+
     include_cross_pols: InitVar[bool] = True
+
+    basis_vector_type = None
 
     # In the future, this might allow for cartesian basis vectors in some orientation.
     # In that case, the Naxes_vec would be 3 rather than 2
@@ -70,6 +71,26 @@ class AnalyticBeam(ABC):
 
     def __init_subclass__(cls):
         """Initialize any imported subclass."""
+        if (
+            cls.__name__ != "UnpolarizedAnalyticBeam"
+            and not hasattr(cls, "_efield_eval")
+            and not hasattr(cls, "_power_eval")
+        ):
+            raise TypeError(
+                "Either _efield_eval or _power_eval method must be defined on "
+                f"{cls.__name__}. Defining _efield_eval is the most general "
+                "approach because it can represent polarized and negative going "
+                "beams. If only _power_eval is defined, the E-field beam is "
+                "defined as the square root of the auto pol power beam."
+            )
+
+        if cls.basis_vector_type is None:
+            warnings.warn(
+                "basis_vector_type was not defined, defaulting to azimuth and "
+                "zenith_angle."
+            )
+            cls.basis_vector_type = "az_za"
+
         if cls.basis_vector_type not in cls._basis_vec_dict:
             raise ValueError(
                 f"basis_vector_type for {cls.__name__} is {cls.basis_vector_type}, "
@@ -77,11 +98,6 @@ class AnalyticBeam(ABC):
             )
 
         cls.__types__[cls.__name__] = cls
-
-    @property
-    @abstractmethod
-    def basis_vector_type(self):
-        """Require that a basis_vector_type is defined in concrete classes."""
 
     def __post_init__(self, include_cross_pols):
         """
@@ -94,20 +110,29 @@ class AnalyticBeam(ABC):
             for the power beam.
 
         """
-        if self.feed_array is not None:
-            for feed in self.feed_array:
-                allowed_feeds = ["n", "e", "x", "y", "r", "l"]
+        if self.feeds is not None:
+            allowed_feeds = ["n", "e", "x", "y", "r", "l"]
[…]
