+++ tests/test_analytic_beam.py
@@ -10,7 +10,8 @@ from astropy.constants import c as speed_of_light
 from scipy.special import j1
 
 from pyuvdata import AiryBeam, GaussianBeam, ShortDipoleBeam, UniformBeam, UVBeam
-from pyuvdata.analytic_beam import AnalyticBeam
+from pyuvdata.analytic_beam import AnalyticBeam, UnpolarizedAnalyticBeam
+from pyuvdata.testing import check_warnings
 
 
 def test_airy_beam_values(az_za_deg_grid):
@@ -157,11 +158,11 @@ def test_diameter_to_sigma(az_za_deg_grid):
 
     wavelengths = speed_of_light.to("m/s").value / freqs
 
-    airy_vals = abm._power_eval(
+    airy_vals = abm.power_eval(
         az_array=az_array.flatten(), za_array=za_array.flatten(), freq_array=freqs
     )
 
-    gauss_vals = gbm._power_eval(
+    gauss_vals = gbm.power_eval(
         az_array=az_array.flatten(), za_array=za_array.flatten(), freq_array=freqs
     )
 
@@ -206,13 +207,27 @@ def test_short_dipole_beam(az_za_deg_grid):
     expected_data[0, 2] = -(np.sin(za_vals) ** 2) * np.sin(2.0 * az_vals) / 2.0
     expected_data[0, 3] = -(np.sin(za_vals) ** 2) * np.sin(2.0 * az_vals) / 2.0
 
-    np.testing.assert_allclose(power_vals, expected_data)
+    np.testing.assert_allclose(power_vals, expected_data, atol=1e-15, rtol=0)
 
-    assert beam.__repr__() == "ShortDipoleBeam(x_orientation='east')"
+    assert beam.__repr__() == "ShortDipoleBeam(feeds=['e', 'n'], x_orientation='east')"
 
 
-def test_uniform_beam(az_za_deg_grid):
-    beam = UniformBeam()
+def test_shortdipole_feed_error():
+    with pytest.raises(
+        ValueError,
+        match=re.escape(
+            "Feeds must be one of: ['n', 'e', 'x', 'y'], got feeds: ['r', 'l']"
+        ),
+    ):
+        ShortDipoleBeam(feeds=["r", "l"])
+
+
+@pytest.mark.parametrize(
+    ("feeds", "x_orientation"),
+    [(None, "east"), (["y", "x"], "north"), (["e", "n"], "east"), (["r", "l"], None)],
+)
+def test_uniform_beam(az_za_deg_grid, feeds, x_orientation):
+    beam = UniformBeam(feeds=feeds, x_orientation=x_orientation)
 
     az_vals, za_vals, freqs = az_za_deg_grid
 
@@ -225,6 +240,13 @@ def test_uniform_beam(az_za_deg_grid):
     np.testing.assert_allclose(beam_vals, expected_data)
 
     assert beam.__repr__() == "UniformBeam()"
+
+    if feeds is not None and "r" in feeds:
+        assert beam.east_ind is None
+        assert beam.north_ind is None
+    else:
+        assert beam.east_ind == 0
+        assert beam.north_ind == 1
 
 
 @pytest.mark.parametrize(
@@ -246,10 +268,11 @@ def test_power_analytic_beam(beam_obj, kwargs, xy_grid):
     efield_vals = beam.efield_eval(az_array=az_vals, za_array=za_vals, freq_array=freqs)
     power_vals = beam.power_eval(az_array=az_vals, za_array=za_vals, freq_array=freqs)
 
-    # check power beams are peak normalized
-    assert np.max(power_vals) == 1.0
-
     atol = 2e-15
+
[…]
