### 2024-10-10 0d0e092
Make defining new analytic beams easier
Add a new Unpolarized beam base class
only require one eval method
add some checking
