+++ docs/analytic_beam_tutorial.rst
@@ -4,10 +4,9 @@
 Analytic Beams
 --------------
 
-The analytic beams defined in pyuvdata are based on an abstract base class,
+The analytic beams defined in pyuvdata are based on a base class,
 :class:`pyuvdata.analytic_beam.AnalyticBeam`, which ensures a standard interface
 and can be used to define other analytic beams in a consistent way.
-
 
 Evaluating analytic beams
 -------------------------
@@ -181,101 +180,125 @@ part, but in general E-Field beams can be complex, so a complex array is returne
 Defining new analytic beams
 ---------------------------
 
-Defining new analytic beams is relatively straight forward. The new beam needs
-to be defined as a `dataclass <https://docs.python.org/3/library/dataclasses.html>`_
-that inherits from :class:`pyuvdata.analytic_beam.AnalyticBeam`, which is an
-abstract base class that specifies what needs to be defined on the new class.
+We have worked to make defining new analytic beams as straight forward as possible.
+The new beam needs inherit from either the :class:`pyuvdata.analytic_beam.AnalyticBeam`,
+or the :class:`pyuvdata.analytic_beam.UnpolarizedAnalyticBeam`, which are base
+classes that specify what needs to be defined on the new class. Unpolarized
+beams (based on the ``UnpolarizedAnalyticBeam`` class) have fewer things that
+need to be specified.
 
-Any parameters that control the beam response (e.g. diameter) must be listed
-with type annotations and optionally defaults to be picked up by the
-dataclass constructor (these are called ``fields`` in the dataclass).
+Note that while unpolarized beams are simpler to define and think about, they
+are quite unphysical and can have results that may be surprising to radio
+astronomers. Since unpolarized feeds respond equally to all orientations of the
+E-field, if two feeds are specified they will have cross-feed power responses that
+are more similar to typical auto-feed power responses (and they will be identical
+to auto-feed power responses if the two feeds have the same beam shapes).
 
-In addition to any fields specific to this new beam, the following fields should
-be specified for polarized beams:
+Setting parameters on the beam
+******************************
 
-  - ``feed_array``: this is an array of feed strings. For polarized beams, this
-    should be specified as required or hardcoded. For example, on the
-    :class:`pyuvdata.ShortDipoleBeam` it is hardcoded as::
+If the new beam has any parameters that control the beam response (e.g. diameter),
+The class must have an ``@dataclass`` decorator and the parameters must be listed
+in the class definitions with type annotations and optionally defaults (these
+are called ``fields`` in the dataclass, see the examples below and
+`dataclass <https://docs.python.org/3/library/dataclasses.html>`_ for more details).
 
-        feed_array = ["e", "n"]
+If you need to do some manipulation or validation of the parameters after they
+are specified by the user, you can use the dataclass's ``__post_init__`` method
+to do that, being sure to call the super class ``__post_init__`` as well.
+The gaussian beam example below shows how this can be done.
 
-  - ``x_orientation``: This species what the ``x`` feed polarization corresponds
-    to, allowed values are ``"east"`` or ``"north"``. For example, on the
-    :class:`pyuvdata.ShortDipoleBeam` it is specified with a default as:
+Polarized beams
+***************
 
-        x_orientation: Literal["east", "north"] = "east"
+For polarized beams (based on the ``AnalyticBeam`` class), the following items
+may be specified, the defaults on the ``AnalyticBeam`` class are noted:
 
+  - ``feeds``: This an list of feed strings. The default is ``["x", "y"]``.
+    This is a a dataclass field, so the the class must have  ``@dataclass``
+    decorator and it should be specied with type annotations and optionally a
[…]
