### 2024-10-10 611ca63
Fix tutorial typos
