### 2024-10-10 67fae2c
go back to `feed_array` for conformity with UVBeam in BeamInterface
