### 2024-10-10 1cbf03d
go back to `feed_array` for conformity with UVBeam in BeamInterface
