+++ src/pyuvdata/analytic_beam.py
@@ -8,7 +8,7 @@ from __future__ import annotations
 import dataclasses
 import importlib
 import warnings
-from dataclasses import InitVar, dataclass, field
+from dataclasses import InitVar, astuple, dataclass, field
 from typing import ClassVar, Literal
 
 import numpy as np
@@ -24,6 +24,29 @@ from .uvbeam.uvbeam import UVBeam, _convert_feeds_to_pols
 __all__ = ["AnalyticBeam", "AiryBeam", "GaussianBeam", "ShortDipoleBeam", "UniformBeam"]
 
 
+def array_safe_eq(a, b) -> bool:
+    """Check if a and b are equal, even if they are numpy arrays."""
+    if a is b:
+        return True
+    if isinstance(a, np.ndarray) and isinstance(b, np.ndarray):
+        return a.shape == b.shape and (a == b).all()
+    try:
+        return a == b
+    except TypeError:  # pragma: no cover
+        return NotImplemented
+
+
+def dc_eq(dc1, dc2) -> bool:
+    """Check if two dataclasses which hold numpy arrays are equal."""
+    if dc1 is dc2:
+        return True
+    if dc1.__class__ is not dc2.__class__:
+        return NotImplemented  # better than False
+    t1 = astuple(dc1)
+    t2 = astuple(dc2)
+    return all(array_safe_eq(a1, a2) for a1, a2 in zip(t1, t2, strict=True))
+
+
 @dataclass(kw_only=True)
 class AnalyticBeam:
     """
@@ -31,7 +54,7 @@ class AnalyticBeam:
 
     Attributes
     ----------
-    feeds : list of str
+    feed_array : array-like of str
         Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
         or r & l.
     x_orientation : str
@@ -42,7 +65,7 @@ class AnalyticBeam:
 
     Parameters
     ----------
-    feeds : list of str
+    feed_array : array-like of str
         Feeds to define this beam for, e.g. x & y or n & e (for "north" and "east")
         or r & l. Default is ["x", "y"].
     x_orientation : str
@@ -56,7 +79,7 @@ class AnalyticBeam:
 
     """
 
-    feeds: list[str] | None = None
+    feed_array: npt.ArrayLike[str] | None = None
     x_orientation: Literal["east", "north"] | None = "east"
 
     include_cross_pols: InitVar[bool] = True
@@ -110,20 +133,21 @@ class AnalyticBeam:
             for the power beam.
 
         """
-        if self.feeds is not None:
+        if self.feed_array is not None:
             allowed_feeds = ["n", "e", "x", "y", "r", "l"]
-            for feed in self.feeds:
+            for feed in self.feed_array:
                 if feed not in allowed_feeds:
                     raise ValueError(
                         f"Feeds must be one of: {allowed_feeds}, "
-                        f"got feeds: {self.feeds}"
+                        f"got feeds: {self.feed_array}"
                     )
+            self.feed_array = np.asarray(self.feed_array)
         else:
-            self.feeds = ["x", "y"]
+            self.feed_array = np.asarray(["x", "y"])
 
         linear_pol = False
         for feed_name in ["x", "y", "e", "n"]:
-            if feed_name in self.feeds:
+            if feed_name in self.feed_array:
                 linear_pol = True
 
         if self.x_orientation is None and linear_pol:
@@ -132,8 +156,14 @@ class AnalyticBeam:
             )
 
         self.polarization_array, _ = _convert_feeds_to_pols(
-            self.feeds, include_cross_pols, x_orientation=self.x_orientation
+            self.feed_array, include_cross_pols, x_orientation=self.x_orientation
         )
+
+    def __eq__(self, other):
+        """Define equality."""
[…]
