### 2024-10-11 #7161 @JBorrow
Hi @ptiede and @pritchardn, I found your name in the JOSS reviewer's database and think you'd be a great reviewer for this package. Would that be something you'd be interested in? Our reviewer's guide is available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
