### 2024-10-11 9d44828
fix docs typos
don't use np.float
