+++ src/pyuvdata/uvbeam/initializers.py
@@ -18,7 +18,7 @@ def new_uvbeam(
     *,
     telescope_name: str,
     data_normalization: Literal["physical", "peak", "solid_angle"],
-    freq_array: npt.NDArray[np.float],
+    freq_array: npt.NDArray[float],
     feed_name: str = "default",
     feed_version: str = "0.0",
     model_name: str = "default",
@@ -31,19 +31,19 @@ def new_uvbeam(
     pixel_coordinate_system: (
         Literal["az_za", "orthoslant_zenith", "healpix"] | None
     ) = None,
-    axis1_array: npt.NDArray[np.float] | None = None,
-    axis2_array: npt.NDArray[np.float] | None = None,
+    axis1_array: npt.NDArray[float] | None = None,
+    axis2_array: npt.NDArray[float] | None = None,
     nside: int | None = None,
     ordering: Literal["ring", "nested"] | None = None,
-    healpix_pixel_array: npt.NDArray[np.int] | None = None,
-    basis_vector_array: npt.NDArray[np.float] | None = None,
-    bandpass_array: npt.NDArray[np.float] | None = None,
-    element_location_array: npt.NDArray[np.float] | None = None,
+    healpix_pixel_array: npt.NDArray[int] | None = None,
+    basis_vector_array: npt.NDArray[float] | None = None,
+    bandpass_array: npt.NDArray[float] | None = None,
+    element_location_array: npt.NDArray[float] | None = None,
     element_coordinate_system: Literal["n-e", "x-y"] | None = None,
-    delay_array: npt.NDArray[np.float] | None = None,
-    gain_array: npt.NDArray[np.float] | None = None,
-    coupling_matrix: npt.NDArray[np.float] | None = None,
-    data_array: npt.NDArray[np.float] | None = None,
+    delay_array: npt.NDArray[float] | None = None,
+    gain_array: npt.NDArray[float] | None = None,
+    coupling_matrix: npt.NDArray[float] | None = None,
+    data_array: npt.NDArray[float] | None = None,
     history: str = "",
     **kwargs,
 ):
