### 2024-10-11 4a67d6d
fix docs typos
don't use np.float
