+++ docs/analytic_beam_tutorial.rst
@@ -216,17 +216,17 @@ may be specified, the defaults on the ``AnalyticBeam`` class are noted:
 
   - ``feed_array``: This an array of feed strings (a list can also be passed,
     it will be converted to an array). The default is ``["x", "y"]``.
-    This is a a dataclass field, so the the class must have  ``@dataclass``
-    decorator and it should be specified with type annotations and optionally a
-    default (see examples below).
+    This is a a dataclass field, so if it is specified, the class must have
+    ``@dataclass`` decorator and it should be specified with type annotations
+    and optionally a default (see examples below).
 
   - ``x_orientation``: For linear polarization feeds, this specifies what the
     ``x`` feed polarization correspond to, allowed values are ``"east"`` or
     ``"north"``, the default is ``"east"``. Should be set to ``None`` for
     circularly polarized feeds.
-    This is a a dataclass field, so the the class must have  ``@dataclass``
-    decorator and it should be specified with type annotations and optionally a
-    default (see examples below).
+    This is a a dataclass field, so if it is specified, the class must have
+    ``@dataclass`` decorator and it should be specified with type annotations
+    and optionally a default (see examples below).
 
   - ``basis_vector_type``: This defines the coordinate system for the
     polarization basis vectors, the default is ``"az_za"``. Currently only
@@ -240,11 +240,11 @@ Defining the beam response
 At least one of the ``_efield_eval`` or ``_power_eval`` methods must be
 defined to specify the response of the new beam. Defining ``_efield_eval`` is
 the most general approach because it can represent complex and negative going
-E-field beams (in this case, power beams will be calculated from the E-field beams).
-If only ``_power_eval`` is defined, the E-field beam is defined as the square
-root of the auto polarization power beam (so will be real and positive definite).
-Both methods can be specified, which may allow for computational efficiencies in
-some cases.
+E-field beams (if only ``_efield_eval`` defined, power beams will be calculated
+from the E-field beams). If only ``_power_eval`` is defined, the E-field beam is
+defined as the square root of the auto polarization power beam, so the E-field
+beam will be real and positive definite. Both methods can be specified, which
+may allow for computational efficiencies in some cases.
 
 The inputs to the ``_efield_eval`` and ``_power_eval`` methods are the same and
 give the directions (azimuth and zenith angle) and frequencies to evaluate the
@@ -438,7 +438,7 @@ Defining a cosine beam with no free parameters is even simpler:
             data_array = self._get_empty_data_array(az_grid.shape, beam_type="power")
 
             for pol_i in np.arange(self.Npols):
-                data_array[0, pol_i, :, :] = np.cos(self.width * za_grid) ** 2
+                data_array[0, pol_i, :, :] = np.cos(za_grid) ** 2
 
             return data_array
 
@@ -539,6 +539,129 @@ method was not defined (we have tests verifying this).
                 data_array[0, 3] = data_array[0, 2]
 
             return data_array
+
+If we wanted to specify the default feed_array to be ``["e", "n"]`` and that the
+default x_orientation was ``"north"`` we would define it as shown below. We
+handle the defaulting of the feed_array in the ``__post_init__`` because dataclass
+fields cannot have mutable defaults. We also do some validation in the
+``__post_init__``.  Note that we call the ``super().__post_init__`` within that
+method to ensure that all the normal AnalyticBeam setup has been done.
+
+.. code-block:: python
+  :linenos:
+
+    from typing import Literal
+    from dataclasses import dataclass
+
+    import numpy as np
+    import numpy.typing as npt
+    from pyuvdata.analytic_beam import AnalyticBeam
+
+    @dataclass(kw_only=True)
[…]
