### 2024-10-14 8b6ee44
Add support for an easier to use validation method
