### 2024-10-14 eeb0e23
Add support for an easier to use validation method
