### 2024-10-15 bc32df8
adding the basic feko read function
