+++ src/pyuvdata/uvbeam/feko_beam.py
@@ -2,6 +2,7 @@
 # Copyright (c) 2018 Radio Astronomy Software Group
 # Licensed under the 2-clause BSD License
 """Class for reading beam FEKO files."""
+
 import os
 import re
 import warnings
@@ -22,11 +23,12 @@ class FEKOBeam(UVBeam):
     read_feko_beam method on the UVBeam class.
 
     """
+
     def nametopol(self, fname):
         """
         Get name of the y file  from the main filename.
 
-    
+
         Parameters
         ----------
         fname : str
@@ -37,10 +39,9 @@ class FEKOBeam(UVBeam):
         str
             New file name.
         """
-        fnew = fname.replace("x","y")
-        
-        return fnew
+        fnew = fname.replace("x", "y")
 
+        return fnew
 
     def read_feko_beam(
         self,
@@ -156,13 +157,12 @@ class FEKOBeam(UVBeam):
         else:
             self.Naxes_vec = 2
             self.Ncomponents_vec = 2
-             
-            
+
             if feed_pol == "x":
-                self.feed_array = np.array(["x" , "y"])
+                self.feed_array = np.array(["x", "y"])
             elif feed_pol == "y":
-                self.feed_array = np.array(["y" , "x"])
-            
+                self.feed_array = np.array(["y", "x"])
+
             self.Nfeeds = self.feed_array.size
             self._set_efield()
 
@@ -184,7 +184,9 @@ class FEKOBeam(UVBeam):
         phi_col = np.where(np.array(column_names) == "Phi")[0][0]
 
         with open(filename, "r") as fh:
-            data_chunks = fh.read()[1:].split(
+            data_chunks = fh.read()[
+                1:
+            ].split(
                 "\n\n"
             )  ## avoiding the first row since there is a blank row at the start of every file
         data_all = [
@@ -193,46 +195,67 @@ class FEKOBeam(UVBeam):
 
         if beam_type == "efield":
             filename2 = self.nametopol(filename)
-            with open(filename2, 'r') as fh:
-                data_chunks = fh.read()[1:].split('\n\n') ## avoiding the first row since there is a blank row at the start of every file
-            data_all2 = [i.splitlines()[9:] for i in data_chunks] ## skips the 9 lines of text in each chunk
+            with open(filename2, "r") as fh:
+                data_chunks = fh.read()[
+                    1:
+                ].split(
+                    "\n\n"
+                )  ## avoiding the first row since there is a blank row at the start of every file
+            data_all2 = [
+                i.splitlines()[9:] for i in data_chunks
+            ]  ## skips the 9 lines of text in each chunk
 
-        frequency = [float(i.split('Frequency')[1].split()[1]) for i in data_chunks[:-1]]
+        frequency = [
+            float(i.split("Frequency")[1].split()[1]) for i in data_chunks[:-1]
+        ]
         self.Nfreqs = len(frequency)
         self.freq_array = np.zeros((1, self.Nfreqs))
         self.freq_array[0] = frequency
         self.bandpass_array = np.zeros((1, self.Nfreqs))
 
-        data_each=np.zeros((len(self.freq_array[0]),np.shape(data_all[0])[0],9))
+        data_each = np.zeros((len(self.freq_array[0]), np.shape(data_all[0])[0], 9))
         if beam_type == "efield":
-            data_each2=np.zeros((len(self.freq_array[0]),np.shape(data_all2[0])[0],9))
-        
+            data_each2 = np.zeros(
+                (len(self.freq_array[0]), np.shape(data_all2[0])[0], 9)
+            )
 
         for i in range(len(self.freq_array[0])):
-
-            data_each[i,:,:] = np.array([list(map(float,data.split())) for data in data_all[i]])
[…]
