+++ src/pyuvdata/uvbeam/uvbeam.py
@@ -4056,7 +4056,7 @@ class UVBeam(UVBase):
         """
         from . import feko_beam
 
-        if not isinstance(filename, (list, tuple)) and filename.endswith("yaml"):
+        if not isinstance(filename, list | tuple) and filename.endswith("yaml"):
             settings_dict = self._read_feko_beam_yaml(filename)
             if not isinstance(settings_dict["frequencies"], list):
                 raise ValueError("frequencies in yaml file must be a list.")
@@ -4082,8 +4082,8 @@ class UVBeam(UVBase):
             for key, val in overriding_keywords.items():
                 if val is not None:
                     warnings.warn(
-                        "The {key} keyword is set, overriding the "
-                        "value in the settings yaml file.".format(key=key)
+                        f"The {key} keyword is set, overriding the "
+                        "value in the settings yaml file."
                     )
 
             if feed_pol is None:
@@ -4131,7 +4131,7 @@ class UVBeam(UVBase):
             rename_extra_keys_map = {"sim_beam_type": "sim_type"}
             for key, value in settings_dict.items():
                 if key not in known_keys:
-                    if key in rename_extra_keys_map.keys():
+                    if key in rename_extra_keys_map:
                         extra_keywords[rename_extra_keys_map[key]] = value
                     else:
                         extra_keywords[key] = value
@@ -4155,9 +4155,9 @@ class UVBeam(UVBase):
                         raise ValueError(f"frequency {freq} not in frequency list")
                 freq_inds = np.array(freq_inds)
                 frequency = freq_array[freq_inds].tolist()
-                cst_filename = np.array(cst_filename)[freq_inds].tolist()
-                if len(cst_filename) == 1:
-                    cst_filename = cst_filename[0]
+                feko_filename = np.array(feko_filename)[freq_inds].tolist()
+                if len(feko_filename) == 1:
+                    feko_filename = feko_filename[0]
                 if isinstance(feed_pol, list):
                     if rotate_pol is None:
                         # if a mix of feed pols, don't rotate by default
@@ -4182,21 +4182,19 @@ class UVBeam(UVBase):
             if len(frequency.shape) > 1:
                 raise ValueError("frequency can not be a multi-dimensional array")
             frequency = frequency.tolist()
-        if isinstance(frequency, (list, tuple)):
-            if len(frequency) == 1:
-                frequency = frequency[0]
+        if isinstance(frequency, list | tuple) and len(frequency) == 1:
+            frequency = frequency[0]
 
         if isinstance(feed_pol, np.ndarray):
             if len(feed_pol.shape) > 1:
                 raise ValueError("feed_pol can not be a multi-dimensional array")
             feed_pol = feed_pol.tolist()
-        if isinstance(feed_pol, (list, tuple)):
-            if len(feed_pol) == 1:
-                feed_pol = feed_pol[0]
+        if isinstance(feed_pol, list | tuple) and len(feed_pol) == 1:
+            feed_pol = feed_pol[0]
 
-        if isinstance(feko_filename, (list, tuple)):
+        if isinstance(feko_filename, list | tuple):
             if frequency is not None:
-                if isinstance(frequency, (list, tuple)):
+                if isinstance(frequency, list | tuple):
                     if not len(frequency) == len(feko_filename):
                         raise ValueError(
                             "If frequency and filename are both "
@@ -4208,7 +4206,7 @@ class UVBeam(UVBase):
             else:
                 freq = None
 
-            if isinstance(feed_pol, (list, tuple)):
+            if isinstance(feed_pol, list | tuple):
                 if not len(feed_pol) == len(feko_filename):
                     raise ValueError(
                         "If feed_pol and filename are both "
[…]
