### 2024-10-17 65d267f
remove mamba lines from benchmarking
