### 2024-10-17 af9077b
Add an error if parameters are not properly passed to flag_init method
