### 2024-10-17 4a65b56
Add an error if parameters are not properly passed to flag_init method
