### 2024-10-21 1f59321
perf: faster UVData.new for rectangular blts
