+++ src/pyuvdata/uvdata/uvdata.py
@@ -1556,7 +1556,8 @@ class UVData(UVBase):
         else:
             app_ra = np.zeros(self.Nblts, dtype=float)
             app_dec = np.zeros(self.Nblts, dtype=float)
-            for cat_id in np.unique(self.phase_center_id_array):
+            cat_ids = np.unique(self.phase_center_id_array)
+            for cat_id in cat_ids:
                 temp_dict = self.phase_center_catalog[cat_id]
                 select_mask = self.phase_center_id_array == cat_id
                 cat_type = temp_dict["cat_type"]
@@ -1570,21 +1571,52 @@ class UVData(UVBase):
                 vrad = temp_dict.get("vrad")
                 dist = temp_dict.get("cat_dist")
 
-                app_ra[select_mask], app_dec[select_mask] = phs_utils.calc_app_coords(
-                    lon_coord=lon_val,
-                    lat_coord=lat_val,
-                    coord_frame=frame,
-                    coord_epoch=epoch,
-                    coord_times=time_val,
-                    pm_ra=pm_ra,
-                    pm_dec=pm_dec,
-                    vrad=vrad,
-                    dist=dist,
-                    time_array=self.time_array[select_mask],
-                    lst_array=self.lst_array[select_mask],
-                    telescope_loc=self.telescope.location,
-                    coord_type=cat_type,
-                )
+                if self.blts_are_rectangular and len(cat_ids) == 1:
+                    slc = (
+                        slice(None, self.Ntimes)
+                        if self.time_axis_faster_than_bls
+                        else slice(None, None, self.Nbls)
+                    )
+                    _app_ra, _app_dec = phs_utils.calc_app_coords(
+                        lon_coord=lon_val,
+                        lat_coord=lat_val,
+                        coord_frame=frame,
+                        coord_epoch=epoch,
+                        coord_times=time_val,
+                        pm_ra=pm_ra,
+                        pm_dec=pm_dec,
+                        vrad=vrad,
+                        dist=dist,
+                        time_array=self.time_array[slc],
+                        lst_array=self.lst_array[slc],
+                        telescope_loc=self.telescope.location,
+                        coord_type=cat_type,
+                        all_times_unique=True,
+                    )
+                    if self.time_axis_faster_than_bls:
+                        app_ra = np.tile(_app_ra, self.Nbls)
+                        app_dec = np.tile(_app_dec, self.Nbls)
+                    else:
+                        app_ra = np.repeat(_app_ra, self.Nbls)
+                        app_dec = np.repeat(_app_dec, self.Nbls)
+                else:
+                    app_ra[select_mask], app_dec[select_mask] = (
+                        phs_utils.calc_app_coords(
+                            lon_coord=lon_val,
+                            lat_coord=lat_val,
+                            coord_frame=frame,
+                            coord_epoch=epoch,
+                            coord_times=time_val,
+                            pm_ra=pm_ra,
+                            pm_dec=pm_dec,
+                            vrad=vrad,
+                            dist=dist,
+                            time_array=self.time_array[select_mask],
+                            lst_array=self.lst_array[select_mask],
+                            telescope_loc=self.telescope.location,
+                            coord_type=cat_type,
+                        )
+                    )
 
         # Now that we have the apparent coordinates sorted out, we can figure
         # out what it is we want to do with the position angle
@@ -4954,19 +4986,44 @@ class UVData(UVBase):
         """
         unprojected_blts = self._check_for_cat_type("unprojected")
 
-        new_uvw = phs_utils.calc_uvw(
-            app_ra=self.phase_center_app_ra,
[…]
