### 2024-10-21 4381bdb
perf: faster UVData.new for rectangular blts
