### 2024-10-21 62ff947
perf: allow specifying checks
