### 2024-10-21 8784862
perf: faster setting of Nants_data in rectangular case
