### 2024-10-21 76974d9
fix: use np.unique instead of set
