### 2024-10-21 #7161 @ptiede
Hi @JBorrow I can review this. Just one thing to note: I am at the same institute as @kartographer, but I haven't worked with him on pyuvfits.
