### 2024-10-21 b9e9a79
Fix a pytest warning
