### 2024-10-21 e41e47d
Remove MANIFEST.in because it's no longer used
setuptools_scm ensures that everything under version control is included in the package
