+++ MANIFEST.in
@@ -1,13 +1,0 @@
-include *.md
-include LICENSE
-include *.bib
-include pyuvdata/uvdata/src/*
-include scripts/*.pro
-include scripts/*.sh
-recursive-include pyuvdata/data *
-recursive-include pyuvdata/*.py
-prune pyuvdata/data/test
-recursive-include docs *
-exclude docs/.DS_Store
-prune docs/_build
-prune docs/__pycache__
