### 2024-10-21 fa869fb
minor docstring improvement
