### 2024-10-22 6ef7f01
Fix plotting in uvbeam tutorial
