+++ docs/uvbeam_tutorial.rst
@@ -84,7 +84,6 @@ a) Reading a CST power beam file
   >>> import os
   >>> from pyuvdata import UVBeam
   >>> from pyuvdata.data import DATA_PATH
-  >>> import matplotlib.pyplot as plt # doctest: +SKIP
 
   # you can pass several filenames and the objects from each file will be
   # combined across the appropriate axis -- in this case frequency
@@ -130,12 +129,6 @@ a) Reading a CST power beam file
   >>> print(beam.data_array.shape)
   (1, 2, 2, 181, 360)
 
-  >>> # plot zenith angle cut through beam
-  >>> plt.plot(beam.axis2_array, beam.data_array[0, 0, 0, 0, :, 0]) # doctest: +SKIP
-  >>> plt.xscale('log') # doctest: +SKIP
-  >>> plt.xlabel('Zenith Angle (radians)') # doctest: +SKIP
-  >>> plt.ylabel('Power') # doctest: +SKIP
-  >>> plt.show() # doctest: +SKIP
 
 b) Reading a CST E-field beam file
 **********************************
@@ -297,10 +290,10 @@ a) Selecting a range of Zenith Angles
   >>> _ = ax.set_ylabel('Power')
   >>> _ = fig.legend(loc="upper right", bbox_to_anchor=[0.9,0.88])
   >>> plt.show()  # doctest: +SKIP
-  >>> plt.savefig("Images/beam_cut.png", bbox_inches='tight')
+  >>> plt.savefig("Images/select_beam_cut.png", bbox_inches='tight')
   >>> plt.clf()
 
-.. image:: Images/beam_cut.png
+.. image:: Images/select_beam_cut.png
   :width: 600
 
 a) Selecting Feeds or Polarizations
@@ -401,23 +394,55 @@ extra interpolation errors.
   >>> import os
   >>> import numpy as np
   >>> from astropy_healpix import HEALPix
-  >>> import matplotlib.pyplot as plt # doctest: +SKIP
-  >>> from matplotlib.colors import LogNorm # doctest: +SKIP
+  >>> import matplotlib.pyplot as plt
+  >>> from matplotlib.colors import LogNorm
   >>> from pyuvdata import UVBeam
   >>> from pyuvdata.data import DATA_PATH
+  >>> import pyuvdata.utils as uvutils
   >>> settings_file = os.path.join(DATA_PATH, 'NicCSTbeams/NicCSTbeams.yaml')
   >>> beam = UVBeam.from_file(settings_file, beam_type='power')
 
-  >>> # this beam file is very large. Let's cut down the size to ease the computation
-  >>> za_max = np.deg2rad(10.0)
+  >>> # Let's cut down to a small area near zenith so we can see the pixelization
+  >>> za_max = np.deg2rad(20.0)
   >>> za_inds_use = np.nonzero(beam.axis2_array <= za_max)[0]
   >>> beam.select(axis2_inds=za_inds_use)
 
   >>> # Optionally specify which interpolation function to use.
   >>> hpx_beam = beam.to_healpix(inplace=False, interpolation_function="az_za_simple")
+
+  >>> # Now plot the pixels on an polar projection with zenith in the center
+  >>> radial_ticks_deg = [5, 10, 20]
+
+  >>> az_array, za_array = np.meshgrid(beam.axis1_array, beam.axis2_array)
[…]
