### 2024-10-22 b1cec23
Update the changelog for the new version
