### 2024-10-22 afd1e39
update minimum pytest version
