+++ ci/pyuvdata_min_deps_tests.yml
@@ -12,7 +12,7 @@ dependencies:
   - scipy>=1.8
   - setuptools_scm>=8.1
   - coverage
-  - pytest>=6.2.5
+  - pytest>=8.2.0
   - pytest-cases>=3.8.3
   - pytest-cov
   - pytest-xdist
