+++ environment.yaml
@@ -16,7 +16,7 @@ dependencies:
   - pre-commit
   - pyerfa>=2.0.1.1
   - pypandoc
-  - pytest>=6.2.5
+  - pytest>=8.2.0
   - pytest-cases>=3.8.3
   - pytest-cov
   - pytest-xdist
