+++ ci/pyuvdata_tests_mac_arm.yml
@@ -15,7 +15,7 @@ dependencies:
   - pyyaml>=5.4.1
   - scipy>=1.8
   - coverage
-  - pytest>=6.2.5
+  - pytest>=8.2.0
   - pytest-cases>=3.8.3
   - pytest-cov
   - pytest-xdist
