### 2024-10-22 da17ef7
update minimum pytest version
