+++ pyproject.toml
@@ -61,7 +61,7 @@ lunar = ["lunarsky>=0.2.4"]
 novas = ["novas", "novas_de405"]
 all = ["pyuvdata[astroquery,casa,hdf5_compression,healpix,lunar,novas]"]
 test = [
-    "pytest>=6.2.5",
+    "pytest>=8.2.0",
     "pytest-xdist",
     "pytest-cases>=3.8.3",
     "pytest-cov",
