### 2024-10-22 87bd9f6
Set this to be version 3.1.0
update a few deprecation warnings that were missing version limits
