### 2024-10-22 1f8e1e8
Set this to be version 3.1.0
update a few deprecation warnings that were missing version limits
