### 2024-10-22 ac85512
increase minimum lunarsky version for compatibility with astropy 6.0
