+++ environment.yaml
@@ -26,6 +26,6 @@ dependencies:
   - setuptools_scm>=8.1
   - sphinx
   - pip:
-      - lunarsky>=0.2.4
+      - lunarsky>=0.2.5
       - novas
       - novas_de405
