+++ ci/pyuvdata_min_versions_tests.yml
@@ -23,6 +23,6 @@ dependencies:
   - setuptools_scm==8.1
   - pip
   - pip:
-      - lunarsky==0.2.4
+      - lunarsky==0.2.5
       - novas
       - novas_de405
