### 2024-10-22 8bbcdc4
increase minimum lunarsky version for compatibility with astropy 6.0
