### 2024-10-24 128036c
Set tolerances in tests as much as possible
Also use `np.testing.assert_allclose` as much as possible
