+++ tests/utils/test_coordinates.py
@@ -415,9 +415,9 @@ def test_lla_xyz_lla_roundtrip():
     lons *= np.pi / 180.0
     xyz = utils.XYZ_from_LatLonAlt(lats, lons, alts)
     lats_new, lons_new, alts_new = utils.LatLonAlt_from_XYZ(xyz)
-    np.testing.assert_allclose(lats_new, lats)
-    np.testing.assert_allclose(lons_new, lons)
-    np.testing.assert_allclose(alts_new, alts)
+    np.testing.assert_allclose(lats_new, lats, rtol=0, atol=utils.RADIAN_TOL)
+    np.testing.assert_allclose(lons_new, lons, rtol=0, atol=utils.RADIAN_TOL)
+    np.testing.assert_allclose(alts_new, alts, rtol=0, atol=1e-3)
 
 
 def test_xyz_from_latlonalt(enu_ecef_info):
@@ -439,7 +439,9 @@ def test_enu_from_ecef(enu_ecef_info):
     enu = utils.ENU_from_ECEF(
         xyz, latitude=center_lat, longitude=center_lon, altitude=center_alt
     )
-    np.testing.assert_allclose(np.stack((east, north, up), axis=1), enu, atol=1e-3)
+    np.testing.assert_allclose(
+        np.stack((east, north, up), axis=1), enu, rtol=0, atol=1e-3
+    )
 
     enu2 = utils.ENU_from_ECEF(
         xyz,
@@ -449,7 +451,7 @@ def test_enu_from_ecef(enu_ecef_info):
             height=center_alt * units.m,
         ),
     )
-    np.testing.assert_allclose(enu, enu2)
+    np.testing.assert_allclose(enu, enu2, rtol=0, atol=1e-3)
 
 
 @pytest.mark.skipif(not hasmoon, reason="lunarsky not installed")
@@ -470,7 +472,9 @@ def test_enu_from_mcmf(enu_mcmf_info, selenoid):
         ellipsoid=selenoid,
     )
 
-    np.testing.assert_allclose(np.stack((east, north, up), axis=1), enu, atol=1e-3)
+    np.testing.assert_allclose(
+        np.stack((east, north, up), axis=1), enu, rtol=0, atol=1e-3
+    )
 
     enu2 = utils.ENU_from_ECEF(
         xyz,
@@ -481,7 +485,7 @@ def test_enu_from_mcmf(enu_mcmf_info, selenoid):
             ellipsoid=selenoid,
         ),
     )
-    np.testing.assert_allclose(enu, enu2, atol=1e-3)
+    np.testing.assert_allclose(enu, enu2, rtol=0, atol=1e-3)
 
 
 def test_invalid_frame():
@@ -612,10 +616,10 @@ def test_ecef_from_enu_roundtrip(enu_ecef_info, enu_mcmf_info, frame, selenoid):
         frame=frame,
         ellipsoid=selenoid,
     )
-    np.testing.assert_allclose(xyz, xyz_from_enu, atol=1e-3)
+    np.testing.assert_allclose(xyz, xyz_from_enu, rtol=0, atol=1e-3)
 
     xyz_from_enu2 = utils.ECEF_from_ENU(enu, center_loc=loc_obj)
-    np.testing.assert_allclose(xyz_from_enu, xyz_from_enu2, atol=1e-3)
+    np.testing.assert_allclose(xyz_from_enu, xyz_from_enu2, rtol=0, atol=1e-3)
 
     if selenoid == "SPHERE":
         enu = utils.ENU_from_ECEF(
@@ -633,7 +637,7 @@ def test_ecef_from_enu_roundtrip(enu_ecef_info, enu_mcmf_info, frame, selenoid):
             altitude=center_alt,
[…]
