### 2024-10-24 532938e
Set tolerances in tests as much as possible
Also use `np.testing.assert_allclose` as much as possible
