+++ tests/uvflag/test_uvflag.py
@@ -665,14 +665,32 @@ def test_hdf5_meta_telescope_location(test_outfile):
     meta = hdf5_utils.HDF5Meta(test_f_file)
     lat, lon, alt = meta.telescope_location_lat_lon_alt
 
-    assert np.isclose(lat, meta.telescope_location_obj.lat.rad)
-    assert np.isclose(lon, meta.telescope_location_obj.lon.rad)
-    assert np.isclose(alt, meta.telescope_location_obj.height.to_value("m"))
+    assert np.isclose(
+        lat, meta.telescope_location_obj.lat.rad, rtol=0, atol=utils.RADIAN_TOL
+    )
+    assert np.isclose(
+        lon, meta.telescope_location_obj.lon.rad, rtol=0, atol=utils.RADIAN_TOL
+    )
+    assert np.isclose(
+        alt, meta.telescope_location_obj.height.to_value("m"), rtol=0, atol=1e-3
+    )
 
     lat_deg, lon_deg, alt = meta.telescope_location_lat_lon_alt_degrees
-    assert np.isclose(lat_deg, meta.telescope_location_obj.lat.deg)
-    assert np.isclose(lon_deg, meta.telescope_location_obj.lon.deg)
-    assert np.isclose(alt, meta.telescope_location_obj.height.to_value("m"))
+    assert np.isclose(
+        lat_deg,
+        meta.telescope_location_obj.lat.deg,
+        rtol=0,
+        atol=np.rad2deg(utils.RADIAN_TOL),
+    )
+    assert np.isclose(
+        lon_deg,
+        meta.telescope_location_obj.lon.deg,
+        rtol=0,
+        atol=np.rad2deg(utils.RADIAN_TOL),
+    )
+    assert np.isclose(
+        alt, meta.telescope_location_obj.height.to_value("m"), rtol=0, atol=1e-3
+    )
 
     if hasmoon:
         from lunarsky import MoonLocation
@@ -780,9 +798,9 @@ def test_read_write_loop_spw(uvdata_obj, test_outfile, telescope_frame, selenoid
     meta = hdf5_utils.HDF5Meta(test_outfile)
     loc_obj = meta.telescope_location_obj
 
-    assert np.isclose(loc_obj.x, uvf.telescope.location.x)
-    assert np.isclose(loc_obj.y, uvf.telescope.location.y)
-    assert np.isclose(loc_obj.z, uvf.telescope.location.z)
+    assert np.isclose(loc_obj.x, uvf.telescope.location.x, rtol=0, atol=1e-3)
+    assert np.isclose(loc_obj.y, uvf.telescope.location.y, rtol=0, atol=1e-3)
+    assert np.isclose(loc_obj.z, uvf.telescope.location.z, rtol=0, atol=1e-3)
 
 
 def test_read_write_loop_missing_shapes(uvdata_obj, test_outfile):
@@ -2370,7 +2388,10 @@ def test_to_baseline_metric(uvdata_obj):
     nt1 = len(ind)
     assert np.all(uvf.metric_array[ind, 15, 0] == 2.1)
     assert np.isclose(
-        uvf.metric_array.mean(), (3.2 * nt0 + 2.1 * nt1) / uvf.metric_array.size
+        uvf.metric_array.mean(),
+        (3.2 * nt0 + 2.1 * nt1) / uvf.metric_array.size,
+        rtol=uvf._metric_array.tols[0],
+        atol=uvf._metric_array.tols[1],
     )
 
 
@@ -2475,7 +2496,7 @@ def test_to_baseline_from_antenna(uvdata_obj, uvf_from_uvcal):
     uvf.check()
 
     uvf2.select(bls=old_baseline, times=old_times)
-    assert np.allclose(or_flags, uvf2.flag_array)
[…]
