### 2024-10-24 3db6fbb
Adding keyword to suppress warnings in _add_phase_center
