### 2024-10-24 2fa970a
Fixing minor issue where we've got multiple (duplicate) coords and multiple (duplicate) times for calc_app_coords
