### 2024-10-24 768de74
Fixing minor issue where we've got multiple (duplicate) coords and multiple (duplicate) times for calc_app_coords
