### 2024-10-24 6d6d7af
Fixing minor issue where we've got multiple (duplicate) coords and multiple (duplicate) times for calc_app_coords
