### 2024-10-24 4823ba4
Fixing minor issue where we've got multiple (duplicate) coords and multiple (duplicate) times for calc_app_coords
