### 2024-10-24 b4dd6d5
Temporarily removing POINTING table write for MS
