### 2024-10-24 0af584d
Temporarily removing POINTING table write for MS
