### 2024-10-24 cf910ac
Temporarily removing POINTING table write for MS
