### 2024-10-24 70a0aff
Temporarily removing POINTING table write for MS
