### 2024-10-24 611cf65
Temporarily removing POINTING table write for MS
