### 2024-10-24 775b269
Temporarily removing POINTING table write for MS
