### 2024-10-24 e0f580e
Temporarily removing POINTING table write for MS
