### 2024-10-24 5bf0ab3
Adding handling for OTF magic
