### 2024-10-24 b486265
Adding handling for OTF magic
