### 2024-10-24 fd29082
Adding handling for OTF magic
