+++ src/pyuvdata/uvdata/mir.py
@@ -7,8 +7,9 @@ import contextlib
 import os
 import warnings
 
+import astropy.units as u
 import numpy as np
-from astropy.coordinates import angular_separation
+from astropy.coordinates import SkyCoord, angular_separation
 from astropy.time import Time
 from docstring_parser import DocstringStyle
 
@@ -17,7 +18,7 @@ from ..docstrings import copy_replace_short_description
 from ..telescopes import known_telescope_location
 from . import UVData, mir_parser
 
-__all__ = ["generate_sma_antpos_dict", "Mir"]
+__all__ = ["Mir", "generate_sma_antpos_dict"]
 
 
 def generate_sma_antpos_dict(filepath):
@@ -598,7 +599,20 @@ class Mir(UVData):
         # Note that these are entries that vary per-integration record (inhid), but
         # we include them here so that we can easily expand the per-integration data
         # to the per-baseline-time length arrays that UVData expects.
-        in_to_blt = ["lst", "mjd", "ara", "adec", "isource", "rinteg"]
+        in_to_blt = [
+            "lst",
+            "mjd",
+            "ara",
+            "adec",
+            "isource",
+            "rinteg",
+            "rar",
+            "decr",
+            "offx",
+            "offy",
+            "inhid",
+            "obsmode",
+        ]
         blt_temp_dict = {}
         suppress_warning = False
         for idx, bl_rec in enumerate(mir_data.bl_data):
@@ -638,6 +652,12 @@ class Mir(UVData):
         ant_2_array = np.zeros(Nblts, dtype=int)
         uvw_array = np.zeros((Nblts, 3), dtype=float)
         phase_center_id_array = np.zeros(Nblts, dtype=int)
+        cat_ra = np.zeros(Nblts, dtype=float)
+        cat_dec = np.zeros(Nblts, dtype=float)
+        off_ra = np.zeros(Nblts, dtype=float)
+        off_dec = np.zeros(Nblts, dtype=float)
+        obs_mode = np.zeros(Nblts, dtype=float)
+        inhid_array = np.zeros(Nblts, dtype=float)
         app_ra = np.zeros(Nblts, dtype=float)
         app_dec = np.zeros(Nblts, dtype=float)
 
@@ -656,6 +676,12 @@ class Mir(UVData):
             phase_center_id_array[blt_key] = temp_dict["isource"]
             app_ra[blt_key] = temp_dict["ara"]
             app_dec[blt_key] = temp_dict["adec"]
+            cat_ra[blt_key] = temp_dict["rar"]
+            cat_dec[blt_key] = temp_dict["decr"]
+            off_ra[blt_key] = temp_dict["offx"]
+            off_dec[blt_key] = temp_dict["offy"]
+            inhid_array[blt_key] = temp_dict["inhid"]
+            obs_mode[blt_key] = temp_dict["obsmode"]
 
         # Finally, assign arrays to attributed
         self.ant_1_array = ant_1_array
@@ -703,10 +729,9 @@ class Mir(UVData):
         isource = np.unique(mir_data.in_data["isource"])
         for sou_id in isource:
             source_mask = mir_data.in_data["isource"] == sou_id
-            source_ra = mir_data.in_data["rar"][source_mask].astype(float)
-            source_dec = mir_data.in_data["decr"][source_mask].astype(float)
             source_epoch = np.mean(mir_data.in_data["epoch"][source_mask]).astype(float)
             source_name = mir_data.codes_data["source"][sou_id]
+
             if source_epoch != 2000.0:
                 # When fed a non-J2000 coordinate, we want to convert that so that it
                 # can easily be written into CASA MS format. In this case, we take the
@@ -720,6 +745,13 @@ class Mir(UVData):
[…]
