### 2024-10-24 0fb73a8
Adding handling for OTF magic
