### 2024-10-24 7282eec
Fixing test tolerances in test_mir_parser
