### 2024-10-24 34cef9d
More minor test clean-up
