### 2024-10-28 58827c2
docs: add docstrings for return_basis_vector
