### 2024-10-28 4897d37
docs: add docstrings for return_basis_vector
