### 2024-10-29 bad390c
test: update test invocations of interp to use return_basis_vector
