### 2024-10-29 73b3bab
test: update test invocations of interp to use return_basis_vector
