### 2024-10-29 e9c95de
test: update test invocations of interp to use return_basis_vector
