### 2024-10-29 5e7c2f2
test: fix broken tests
