### 2024-10-29 1b4a030
docs: add docstrings for check_kw and all_times_unique
