### 2024-10-29 fd95c56
fix: force re-set of rectangularity in harmonize_resample_arrays
