### 2024-10-29 ce35398
Address review comments
also fix a couple more places I missed before.
