### 2024-10-29 dca1c0a
Fix a few more places per review comments
