### 2024-10-29 ca4e93a
Fix a few more places per review comments
