### 2024-10-30 c64fa02
docs: fix docstring for return_basis_vector
