### 2024-10-30 12cf3e9
style: use *= instead of /=
