### 2024-10-30 7936c9b
test: update tests to explicitly set return_basis_vector
