### 2024-10-30 bcf73ec
test: update tests to explicitly set return_basis_vector
