+++ tests/uvbeam/test_uvbeam.py
@@ -679,6 +679,7 @@ def test_freq_interpolation(
         return_bandpass=True,
         return_coupling=need_coupling,
         interpolation_function=interpolation_function,
+        return_basis_vector=True,
     )
     if antenna_type == "simple":
         interp_data, interp_basis_vector, interp_bandpass = interp_arrays
@@ -793,7 +794,9 @@ def test_freq_interpolation(
     _pb = beam.select(frequencies=beam.freq_array[:1], inplace=False)
 
     try:
-        interp_data, interp_basis_vector = _pb.interp(freq_array=_pb.freq_array)
+        interp_data, _ = _pb.interp(
+            freq_array=_pb.freq_array, return_basis_vector=False
+        )
     except ValueError as err:
         raise AssertionError(
             "UVBeam.interp didn't return an array slice as expected"
@@ -805,7 +808,9 @@ def test_freq_interpolation(
         ValueError, match="Only one frequency in UVBeam so cannot interpolate."
     ):
         beam_singlef.interp(
-            freq_array=np.array([150e6]), interpolation_function=interpolation_function
+            freq_array=np.array([150e6]),
+            interpolation_function=interpolation_function,
+            return_basis_vector=False,
         )
 
 
@@ -841,11 +846,11 @@ def test_freq_interp_real_and_complex(cst_power_2freq):
 
     # interpolate cubic on real data
     freqs = np.linspace(123e6, 150e6, 10)
-    pb_int = pbeam.interp(freq_array=freqs)[0]
+    pb_int = pbeam.interp(freq_array=freqs, return_basis_vector=False)[0]
 
     # interpolate cubic on complex data and compare to ensure they are the same
     pbeam.data_array = pbeam.data_array.astype(np.complex128)
-    pb_int2 = pbeam.interp(freq_array=freqs)[0]
+    pb_int2 = pbeam.interp(freq_array=freqs, return_basis_vector=False)[0]
     np.testing.assert_allclose(np.abs(pb_int - pb_int2), 0)
 
 
@@ -910,22 +915,24 @@ def test_spatial_interpolation_samepoints(
             "function"
         ),
     ):
-        interp_data_array, interp_basis_vector = uvbeam.interp(
+        interp_data_array, _ = uvbeam.interp(
             az_array=az_orig_vals,
             za_array=za_orig_vals,
             freq_array=freq_orig_vals,
             interpolation_function="healpix_simple",
+            return_basis_vector=False,
         )
 
     # test error if not set to known function
     with pytest.raises(
         ValueError, match="interpolation_function not recognized, must be one of "
     ):
-        interp_data_array, interp_basis_vector = uvbeam.interp(
+        interp_data_array, _ = uvbeam.interp(
             az_array=az_orig_vals,
             za_array=za_orig_vals,
             freq_array=freq_orig_vals,
             interpolation_function="foo",
+            return_basis_vector=False,
         )
 
     # test that new object from interpolation is identical
@@ -975,6 +982,7 @@ def test_spatial_interpolation_samepoints(
             freq_array=freq_orig_vals,
             new_object=True,
             interpolation_function=interpolation_function,
+            return_basis_vector=False,
         )
 
     if beam_type == "power":
@@ -985,6 +993,7 @@ def test_spatial_interpolation_samepoints(
             freq_array=freq_orig_vals,
             polarizations=["xx"],
             interpolation_function=interpolation_function,
+            return_basis_vector=True,
         )
 
         data_array_compare = uvbeam.data_array[:, :1]
@@ -1042,6 +1051,7 @@ def test_spatial_interpolation_everyother(
[…]
