### 2024-10-31 d71b4db
test: add test for deprecation warning
