### 2024-10-31 #7161 @warrickball
Hi @ptiede, I'm taking over some of the assistant editor-in-chief duties and just want to ask for a bit more detail about your collaboration with the authors. In principle your co-authorships and employment at the same institution represent conflicts of interest by the [JOSS policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) but the co-authored papers have long authors lists and the institution you're both employed at happens to be a particularly large one.

Are you confident that you can make an impartial assessment of the `pyuvdata`?
