### 2024-11-05 6149f4c
fix: set include_cross_pols to False if feed_array has length 1
