### 2024-11-08 d124ef1
feat: new convenience methods on BeamInterface
