### 2024-11-08 4a9c6b2
feat: new convenience methods on BeamInterface
