### 2024-11-09 0c7e6ea
test: add tests of with_feeds and as_power_beam
