### 2024-11-11 bd9bcd2
test: don't use healpix coords to go to uvbeam
