+++ tests/test_beam_interface.py
@@ -11,9 +11,20 @@ from pyuvdata import (
     GaussianBeam,
     ShortDipoleBeam,
     UniformBeam,
+    UVBeam,
     utils,
 )
 from pyuvdata.testing import check_warnings
+
+
+@pytest.fixture(scope="function")
+def airy() -> AiryBeam:
+    return AiryBeam(diameter=14.0)
+
+
+@pytest.fixture()
+def gaussian() -> GaussianBeam:
+    return GaussianBeam(diameter=14.0)
 
 
 @pytest.fixture()
@@ -32,6 +43,14 @@ def xy_grid_coarse():
     az_array = np.arctan2(y_arr, x_arr) + np.pi  # convert from -180->180 to 0->360
 
     return az_array, za_array, freqs
+
+
+@pytest.fixture()
+def gaussian_uv(gaussian, az_za_coords) -> UVBeam:
+    az, za = az_za_coords
+    return gaussian.to_uvbeam(
+        axis1_array=az, axis2_array=za, freq_array=np.array([1e8])
+    )
 
 
 @pytest.mark.parametrize(
@@ -245,34 +264,23 @@ def test_compute_response_errors(param, value):
         bi_uvb.compute_response(**compute_kwargs)
 
 
-@pytest.mark.parametrize(
-    "beam_obj",
-    [
-        AiryBeam(diameter=14.0),
-        GaussianBeam(diameter=14.0),
-        AiryBeam(diameter=14.0).to_uvbeam(
-            freq_array=np.array([1e8]), pixel_coordinate_system="healpix", nside=32
-        ),
-    ],
-)
-def test_idempotent_instantiation(beam_obj):
-    beam = BeamInterface(beam_obj)
+@pytest.mark.parametrize("beam_obj", ["airy", "gaussian", "gaussian_uv"])
+def test_idempotent_instantiation(beam_obj, request):
+    beam = BeamInterface(request.getfixturevalue(beam_obj))
     beam2 = BeamInterface(beam)
     assert beam == beam2
 
 
-def test_properties():
-    beam = AiryBeam(diameter=14.0)
-    intf = BeamInterface(beam)
-    assert beam.Npols == intf.Npols
-    assert beam.Nfeeds == intf.Nfeeds
-    assert np.all(beam.polarization_array == intf.polarization_array)
-    assert np.all(beam.feed_array == intf.feed_array)
+def test_properties(airy: AiryBeam):
+    intf = BeamInterface(airy)
+    assert airy.Npols == intf.Npols
+    assert airy.Nfeeds == intf.Nfeeds
+    assert np.all(airy.polarization_array == intf.polarization_array)
+    assert np.all(airy.feed_array == intf.feed_array)
 
 
-def test_clone():
-    beam = AiryBeam(diameter=14.0)
-    intf = BeamInterface(beam)
+def test_clone(airy):
+    intf = BeamInterface(airy)
     intf_clone = intf.clone(beam_type="power")
     assert intf != intf_clone
 
@@ -280,11 +288,10 @@ def test_clone():
 @pytest.mark.parametrize("uvbeam", [True, False], ids=["uvbeam", "analytic"])
 @pytest.mark.parametrize("allow_mutation", [True, False], ids=["mutate", "nomutate"])
 @pytest.mark.parametrize("include_cross_pols", [True, False], ids=["incx", "nox"])
-def test_as_power(uvbeam: bool, allow_mutation: bool, include_cross_pols: bool):
-    beam = AiryBeam(diameter=14.0)
-    if uvbeam:
-        beam = beam.to_uvbeam(freq_array=np.array([1e8]), nside=32)
-
+def test_as_power(
[…]
