### 2024-11-13 4d25cf6
test: fix tests that I broke
