### 2024-11-14 #7161 @JBorrow
Hi @ptiede just checking in to see if you're ok with the JOSS COI policy and can provide an impartial review of `pyuvdata`?
