### 2024-11-14 #7482 @editorialbot
[REVIEW]: pyuvdata v3: an interface for astronomical interferometeric data sets in Python
**Submitting author:** @kartographer (<a href="http://orcid.org/0000-0002-3490-146X">Garrett Keating</a>)
**Repository:** https://github.com/RadioAstronomySoftwareGroup/pyuvdata/
**Branch with paper.md** (empty if default branch): main
**Version:** v3.2.1
