### 2024-11-17 5a23a95
docs: update docstring with info about what the user asked for
