### 2024-11-18 ea18751
Add comments about the in-place corrections
