### 2024-11-20 fe8a3f3
docs: merge changelog sections
