### 2024-11-20 bf1370a
test: fix extraneous warning
