### 2024-11-20 c164317
fix: can't assume that time/baseline ordering means rectangularity
