### 2024-11-20 0806018
fix: can't assume that time/baseline ordering means rectangularity
