### 2024-11-21 2b3eaf3
maint: remove cruft
