### 2024-11-21 60408dc
fix mwa beam docstring
