### 2024-11-21 686af99
update the changelog for v3.1.2
