### 2024-11-21 2324700
Adding warning message to OTF handling
