### 2024-11-21 05aa0ec
Adding warning message to OTF handling
