### 2024-11-22 9cbbbec
Adding ATA to known telescopes
