+++ src/pyuvdata/telescopes.py
@@ -95,6 +95,7 @@ _KNOWN_TELESCOPES = {
             height=1019.222 * units.m,
         ),
         "antenna_diameters": 6.1,
+        "citation": "Private communication (2024)",
     },
 }
 
