### 2024-11-23 561773b
Adding ATA to README telescopes list
