### 2024-12-02 9435eeb
Fix new ruff errors related to string formatting of variables
