### 2024-12-02 6f995a6
Fix new ruff errors related to string formatting of variables
