### 2024-12-03 809e83e
Adding further notes on source of ATA position
