### 2024-12-03 bff7807
cast mir rx_code to an int
