### 2024-12-03 fb44d11
+^ look_in_catalog avoid unrelated keys for cat_type
