### 2024-12-03 020ce0c
Fix bug where polarizations from 3-tuple bls weren't all being kept
