### 2024-12-03 dab9ed7
test: of selecting more than one pol using bls=
