### 2024-12-03 9317875
test: of selecting more than one pol using bls=
