### 2024-12-03 57edc93
*- look_in_catalog minimise fix
