### 2024-12-04 18b6a04
+ tests uvh5 sidereal addition #1506
