### 2024-12-04 4edf829
Add option to be able to turn off writing double precision uvws.
