### 2024-12-04 75fbcb2
Add tests for uvw_double switch
