### 2024-12-08 6789974
+ tests utils look_in_catalog
