### 2024-12-09 2ac1880
+ CHANGELOG look_in_catalog bug fix
