### 2025-02-25 #7482 @David-McKenna
I believe I've finished my review. Apologies for the delay in completing the final points of the checklist, my schedule became a bit more hectic than expected, but the delay ended up allowing me to use `pyuvdata` as a part of my work prior to submitting the review.

I'm not sure that the paper completely hits the "State of the field" section requirements, but given that there isn't much that `pyuvdata` can be compared to other than the large suites of original software used to create / interact with the individual formats it supports, I believe it is in an overall good state.
