### 2025-04-30 #7482 @JBorrow
Hi @kartographer, please can you follow the 'Additional Author Tasks After Review is Complete' above? I will be taking a look at the paper text and those outputs, and then we can proceed to recommending the paper for publication. From there, it goes to the Editor in Chief for the final editorial checks.
