### 2025-04-30 #7482 @JBorrow
Hi @kartographer, a few copy editing notes that you may wish to fix:

- Throughout there is an inconsistent use of the oxford comma (L61 vs L22). You may wish to run a search for 'and' and make sure this is consistent (preferably _including_ the oxford comma).
- 'measurement set' is listed twice in the opening paragraph; is this intended (i.e. it is both a 'visibility file' and a 'calibration solution'?
- The isolated clause (between the two commas) on line 26 doesn't make much sense; you may wish to re-write this line.
- On line 31, 'that typically built to' should be re-worded to make sense. (That 'are' typically built to?)
- When lising the instruments and their URLs on line 37 and 38, please make the URLs clickable and re-format it to avoid nested parentheses - e.g. you could do MWA: http://..., PAPER: http://...,
- On line 55 you use 'i.e.' in a case where it could be an example case (and hence should read 'e.g.' instead); I am not sure though. Please check this.
- There is a rendering error on line 59 where the final bullet point is not listed correctly.
- On line 61 you may wish to isolate the 'averaging in time and frequency' clause by placing a comma after 'frequency'.
