### 2025-05-06 #7482 @JBorrow
Hi @kartographer please make sure the title of your zenodo deposit is the same as the paper (i.e. pyuvdata v3: an interface for astronomical interferometeric datasets in python)
