### 2025-05-06 #7482 @kartographer
@JBorrow - I think I see the issue, following the revisions requested and the release, we merged the `joss_update` branch into our `main` (which should now be the right branch to point at).
