### 2025-05-08 #7482 @warrickball
Hi! I've just opened [a small pull request](https://github.com/RadioAstronomySoftwareGroup/pyuvdata/pull/1560) with some editorial tweaks. You should review those and many any appropriate changes (including possibly just merging mine), after which I'll launch another test run of the publication pipeline.

Note that the changes include tweaks to the title, so you'll need to update the Zenodo archive title to match.
