### 2025-05-12 #7482 @warrickball
Could you tweak the Zenodo archive title to match? I.e., change "datasets in python" to "data sets in Python". Thanks!

I'll re-run the publication pipeline in the meantime anyway to check for any other outstanding issues.
