### 2025-05-13 #7482 @kartographer
@warrickball - apologies for the late notice, but we _just_ noticed that there's a typo in the title of the paper -- the title on Zenodo actually has the correct name: "interferometeric" should be "interferometric". The current version from our main branch has the correct title, is there any way to propagate that change up to JOSS? Apologies again, and thank you in advance!
