### 2025-05-13 #7482 @warrickball
Good spot, I should've seen that too! I'll fix the title of this issue, switch branch here and rebuild the paper.
